requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --output-file=dev.txt dev.in
#
amqp==5.2.0
    # via
    #   -r base.txt
    #   kombu
asgiref==3.7.2
    # via
    #   -r base.txt
    #   django
    #   django-cors-headers
astroid==2.15.8
    # via flake8-django
asttokens==2.4.1
    # via
    #   -r base.txt
    #   stack-data
attrs==23.2.0
    # via flake8-bugbear
billiard==4.2.0
    # via
    #   -r base.txt
    #   celery
black==23.12.1
    # via -r dev.in
celery==5.3.6
    # via -r base.txt
certifi==2023.11.17
    # via
    #   -r base.txt
    #   requests
cffi==1.16.0
    # via
    #   -r base.txt
    #   cryptography
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.3.2
    # via
    #   -r base.txt
    #   requests
click==8.1.7
    # via
    #   -r base.txt
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via
    #   -r base.txt
    #   celery
click-plugins==1.1.1
    # via
    #   -r base.txt
    #   celery
click-repl==0.3.0
    # via
    #   -r base.txt
    #   celery
coverage[toml]==7.4.0
    # via pytest-cov
cryptography==41.0.7
    # via
    #   -r base.txt
    #   social-auth-core
decorator==5.1.1
    # via
    #   -r base.txt
    #   ipython
defusedxml==0.8.0rc2
    # via
    #   -r base.txt
    #   python3-openid
    #   social-auth-core
distlib==0.3.8
    # via virtualenv
django==5.0.1
    # via
    #   -r base.txt
    #   django-admin-display
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-jazzmin
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   social-auth-app-django
django-admin-display==1.3.0
    # via -r base.txt
django-cors-headers==4.3.1
    # via -r base.txt
django-debug-toolbar==4.2.0
    # via -r base.txt
django-extensions==3.2.3
    # via -r base.txt
django-filter==23.5
    # via -r base.txt
django-jazzmin==2.6.0
    # via -r base.txt
django-stubs==4.2.7
    # via -r dev.in
django-stubs-ext==4.2.7
    # via django-stubs
django-templated-mail==1.1.1
    # via
    #   -r base.txt
    #   djoser
djangorestframework==3.14.0
    # via
    #   -r base.txt
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via
    #   -r base.txt
    #   djoser
djoser==2.2.2
    # via -r base.txt
drf-flex-fields==1.0.2
    # via -r base.txt
drf-yasg==1.21.7
    # via -r base.txt
environs==10.3.0
    # via -r base.txt
exceptiongroup==1.2.0
    # via
    #   -r base.txt
    #   ipython
    #   pytest
execnet==2.0.2
    # via pytest-xdist
executing==2.0.1
    # via
    #   -r base.txt
    #   stack-data
factory-boy==3.3.0
    # via -r dev.in
faker==22.4.0
    # via factory-boy
filelock==3.13.1
    # via virtualenv
flake8==6.1.0
    # via
    #   -r dev.in
    #   flake8-bugbear
    #   flake8-comprehensions
    #   flake8-debugger
    #   flake8-django
    #   flake8-use-pathlib
flake8-bugbear==24.1.17
    # via -r dev.in
flake8-comprehensions==3.14.0
    # via -r dev.in
flake8-debugger==4.1.2
    # via -r dev.in
flake8-django==1.4
    # via -r dev.in
flake8-logging-format==0.9.0
    # via -r dev.in
flake8-use-pathlib==0.3.0
    # via -r dev.in
identify==2.5.33
    # via pre-commit
idna==3.6
    # via
    #   -r base.txt
    #   requests
inflection==0.5.1
    # via
    #   -r base.txt
    #   drf-yasg
iniconfig==2.0.0
    # via pytest
ipython==8.20.0
    # via -r base.txt
isort==5.13.2
    # via -r dev.in
jedi==0.19.1
    # via
    #   -r base.txt
    #   ipython
kombu==5.3.5
    # via
    #   -r base.txt
    #   celery
lazy-object-proxy==1.10.0
    # via astroid
marshmallow==3.20.2
    # via
    #   -r base.txt
    #   environs
matchlib==0.2.2
    # via -r dev.in
matplotlib-inline==0.1.6
    # via
    #   -r base.txt
    #   ipython
mccabe==0.7.0
    # via flake8
mypy==1.8.0
    # via -r dev.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
oauthlib==3.2.2
    # via
    #   -r base.txt
    #   requests-oauthlib
    #   social-auth-core
packaging==23.2
    # via
    #   -r base.txt
    #   black
    #   drf-yasg
    #   marshmallow
    #   pytest
    #   pytest-sugar
parso==0.8.3
    # via
    #   -r base.txt
    #   jedi
pathspec==0.12.1
    # via black
pexpect==4.9.0
    # via
    #   -r base.txt
    #   ipython
pillow==10.2.0
    # via -r base.txt
platformdirs==4.1.0
    # via
    #   black
    #   virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.6.0
    # via -r dev.in
prompt-toolkit==3.0.43
    # via
    #   -r base.txt
    #   click-repl
    #   ipython
psycopg2-binary==2.9.9
    # via -r base.txt
ptyprocess==0.7.0
    # via
    #   -r base.txt
    #   pexpect
pure-eval==0.2.2
    # via
    #   -r base.txt
    #   stack-data
pycodestyle==2.11.1
    # via
    #   flake8
    #   flake8-debugger
pycparser==2.21
    # via
    #   -r base.txt
    #   cffi
pyflakes==3.1.0
    # via flake8
pygments==2.17.2
    # via
    #   -r base.txt
    #   ipython
pyjwt==2.8.0
    # via
    #   -r base.txt
    #   djangorestframework-simplejwt
    #   social-auth-core
pytest==7.4.4
    # via
    #   -r dev.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-django
    #   pytest-sugar
    #   pytest-xdist
pytest-asyncio==0.23.3
    # via -r dev.in
pytest-cov==4.1.0
    # via -r dev.in
pytest-django==4.7.0
    # via -r dev.in
pytest-sugar==0.9.7
    # via -r dev.in
pytest-xdist==3.5.0
    # via -r dev.in
python-dateutil==2.8.2
    # via
    #   -r base.txt
    #   celery
    #   faker
    #   time-machine
python-dotenv==1.0.0
    # via
    #   -r base.txt
    #   environs
python3-openid==3.2.0
    # via
    #   -r base.txt
    #   social-auth-core
pytz==2023.3.post1
    # via
    #   -r base.txt
    #   djangorestframework
    #   drf-yasg
pyyaml==6.0.1
    # via
    #   -r base.txt
    #   drf-yasg
    #   pre-commit
requests==2.31.0
    # via
    #   -r base.txt
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via
    #   -r base.txt
    #   social-auth-core
six==1.16.0
    # via
    #   -r base.txt
    #   asttokens
    #   python-dateutil
social-auth-app-django==5.4.0
    # via
    #   -r base.txt
    #   djoser
social-auth-core==4.5.1
    # via
    #   -r base.txt
    #   social-auth-app-django
sqlparse==0.4.4
    # via
    #   -r base.txt
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via
    #   -r base.txt
    #   ipython
termcolor==2.4.0
    # via pytest-sugar
time-machine==2.13.0
    # via -r dev.in
tomli==2.0.1
    # via
    #   black
    #   coverage
    #   django-stubs
    #   mypy
    #   pytest
traitlets==5.14.1
    # via
    #   -r base.txt
    #   ipython
    #   matplotlib-inline
types-pytz==2023.3.1.1
    # via django-stubs
types-pyyaml==6.0.12.12
    # via django-stubs
typing-extensions==4.9.0
    # via
    #   -r base.txt
    #   asgiref
    #   astroid
    #   black
    #   django-stubs
    #   django-stubs-ext
    #   mypy
tzdata==2023.4
    # via
    #   -r base.txt
    #   celery
uritemplate==4.1.1
    # via
    #   -r base.txt
    #   drf-yasg
urllib3==2.1.0
    # via
    #   -r base.txt
    #   requests
vine==5.1.0
    # via
    #   -r base.txt
    #   amqp
    #   celery
    #   kombu
virtualenv==20.25.0
    # via pre-commit
wcwidth==0.2.13
    # via
    #   -r base.txt
    #   prompt-toolkit
wrapt==1.16.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
